tools/ioc-es2mysql/requirements.txt
# IOC ES→MySQL Sync Tool — Python Dependencies
pycti>=6,<7
mysql-connector-python>=8.0,<9.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pyyaml>=6.0,<7.0
geoip2>=4.0,<5.0
